# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoTasks is a robust API for managing company tasks and collaborations.\nIt provides comprehensive features for:\n- **Company and Collaborator Management**: Onboard companies and manage staff with hierarchical roles.\n- **Activity and Task Lifecycle Tracking**: Create and monitor activities and tasks through various stages.\n- **Role-Based Access Control (RBAC)**: Secure endpoints with granular permissions for Admins, Ops, Owners, and Managers.\n- **Performance Ratings and Assignments**: Assign tasks to collaborators and rate their performance.\n- **Secure Authentication**: JWT-based authentication for secure access.",
        "title": "GoTasks API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/company/owner/collaborator/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List collaborators in the authenticated user's company with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collaborator | Company Owner"
                ],
                "summary": "List collaborators of my company",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "role"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Field to sort by",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "default": "ASC",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter expression",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_Paginated-internal_collaborator_CompanyCollaboratorDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new collaborator linked to an existing account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collaborator | Company Owner"
                ],
                "summary": "Create collaborator",
                "parameters": [
                    {
                        "description": "Collaborator create payload",
                        "name": "collaborator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_collaborator.CollaboratorCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_collaborator_CollaboratorInfoDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/company/owner/collaborator/with-account": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new collaborator and account within the company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collaborator | Company Owner"
                ],
                "summary": "Create collaborator with new account",
                "parameters": [
                    {
                        "description": "Collaborator with account create payload",
                        "name": "collaborator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_collaborator.CollaboratorWithAccountCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_collaborator_CompanyCollaboratorDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/company/owner/collaborator/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a company collaborator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collaborator | Company Owner"
                ],
                "summary": "Get collaborator by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Collaborator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_collaborator_CompanyCollaboratorDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a company collaborator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collaborator | Company Owner"
                ],
                "summary": "Update collaborator",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Collaborator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Collaborator update payload",
                        "name": "collaborator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_collaborator.CollaboratorUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-internal_collaborator_CompanyCollaboratorDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a company collaborator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collaborator | Company Owner"
                ],
                "summary": "Delete collaborator",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Collaborator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-tasks_internal_pkg_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-tasks_internal_pkg_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.MessageDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Paginated-internal_collaborator_CompanyCollaboratorDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_collaborator.CompanyCollaboratorDTO"
                    }
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_MessageDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/go-tasks_internal_pkg_httpx.MessageDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-go-tasks_internal_pkg_httpx_Paginated-internal_collaborator_CompanyCollaboratorDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/go-tasks_internal_pkg_httpx.Paginated-internal_collaborator_CompanyCollaboratorDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "go-tasks_internal_pkg_httpx.Response-internal_collaborator_CollaboratorInfoDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_collaborator.CollaboratorInfoDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "go-tasks_internal_pkg_httpx.Response-internal_collaborator_CompanyCollaboratorDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_collaborator.CompanyCollaboratorDTO"
                },
                "error": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_collaborator.CollaboratorAccountCreateDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "owner@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Joe Owner"
                },
                "password": {
                    "type": "string",
                    "example": "S3cureP@ss!"
                },
                "phone": {
                    "type": "string",
                    "example": "+1-202-555-0199"
                },
                "username": {
                    "type": "string",
                    "example": "owner_joe"
                }
            }
        },
        "internal_collaborator.CollaboratorAccountDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "223e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+1-202-555-0123"
                },
                "username": {
                    "type": "string",
                    "example": "jdoe"
                }
            }
        },
        "internal_collaborator.CollaboratorAccountUpdateDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Jane Manager"
                }
            }
        },
...
        "internal_collaborator.CollaboratorCreateDTO": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "account_id": {
                    "type": "string",
                    "example": "223e4567-e89b-12d3-a456-426614174000"
                },
                "role": {
                    "enum": [
                        "MANAGER",
                        "OPS"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_collaborator.ManagedCollaboratorRoleDTO"
                        }
                    ],
                    "example": "MANAGER"
                }
            }
        },
        "internal_collaborator.CollaboratorInfoDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174002"
                },
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "company_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_collaborator.CollaboratorRoleDTO"
                        }
                    ],
                    "example": "OWNER"
                }
            }
        },
...
        "internal_collaborator.CollaboratorRoleDTO": {
            "type": "string",
            "enum": [
                "OWNER",
                "MANAGER",
                "OPS"
            ],
            "x-enum-varnames": [
                "CollaboratorRoleDTOOwner",
                "CollaboratorRoleDTOManager",
                "CollaboratorRoleDTOOps"
            ]
        },
        "internal_collaborator.CollaboratorUpdateDTO": {
            "type": "object",
            "required": [
                "account",
                "role"
            ],
            "properties": {
                "account": {
                    "$ref": "#/definitions/internal_collaborator.CollaboratorAccountUpdateDTO"
                },
                "role": {
                    "enum": [
                        "MANAGER",
                        "OPS"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_collaborator.ManagedCollaboratorRoleDTO"
                        }
                    ],
                    "example": "MANAGER"
                }
            }
        },
        "internal_collaborator.CollaboratorWithAccountCreateDTO": {
            "type": "object",
            "required": [
                "account",
                "role"
            ],
            "properties": {
                "account": {
                    "$ref": "#/definitions/internal_collaborator.CollaboratorAccountCreateDTO"
                },
                "role": {
                    "enum": [
                        "MANAGER",
                        "OPS"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_collaborator.ManagedCollaboratorRoleDTO"
                        }
                    ],
                    "example": "MANAGER"
                }
            }
        },
...
        "internal_collaborator.CompanyCollaboratorDTO": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/internal_collaborator.CollaboratorAccountDTO"
                },
                "account_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174002"
                },
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "company_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_collaborator.CollaboratorRoleDTO"
                        }
                    ],
                    "example": "OWNER"
                }
            }
        },
        "internal_collaborator.ManagedCollaboratorRoleDTO": {
            "type": "string",
            "enum": [
                "MANAGER",
                "OPS"
            ],
            "x-enum-varnames": [
                "ManagedCollaboratorRoleDTOManager",
                "ManagedCollaboratorRoleDTOOps"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}